update for protecting endpoints and authorization

+++ app/app.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI , HTTPException , File , UploadFile , Form , Depends
 from app.schemas import PostCreate ,PostResponse
-from app.db import Post , create_db_and_tables , get_async_session
+from app.db import Post , create_db_and_tables , get_async_session , User
 from sqlalchemy.ext.asyncio import AsyncSession
 from contextlib import asynccontextmanager
 from sqlalchemy import select
@@ -10,7 +10,8 @@ import shutil
 import os
 import uuid
 import tempfile
-
+from app.users import auth_backend , current_active_user , fastapi_users
+from app.schemas import PostCreate, PostResponse , UserRead ,UserCreate ,UserUpdate
 
 @asynccontextmanager
 async def lifespan(app:FastAPI):
@@ -18,11 +19,18 @@ async def lifespan(app:FastAPI):
     yield
 
 app =FastAPI( lifespan=lifespan)
+app.include_router(fastapi_users.get_auth_router(auth_backend), prefix= '/auth/jwt', tags=["auth"])
+app.include_router(fastapi_users.get_register_router(UserRead,UserCreate), prefix= '/auth', tags=["auth"])
+app.include_router(fastapi_users.get_reset_password_router(),prefix='/auth',tags =['auth'])
+app.include_router(fastapi_users.get_verify_router(UserRead),prefix='/auth',tags =['auth'])
+app.include_router(fastapi_users.get_users_router(UserRead, UserUpdate),prefix='/users',tags =['users'])
+
 
 @app.post("/upload")
 async def upload_file(
     file: UploadFile = File(...),
     caption: str = Form(""),
+    user: User =Depends(current_active_user),
     session: AsyncSession = Depends(get_async_session)
 ):
     temp_file_path = None
@@ -44,6 +52,7 @@ async def upload_file(
             
 
             post = Post(
+                user_id =user.id,
                 caption = caption,
                 url = upload_result.url,
                 file_type = "video" if file.content_type.startswith("video/") else "image",
@@ -64,28 +73,36 @@ async def upload_file(
 
 @app.get("/feed")
 async def get_feed(
-    session: AsyncSession =Depends(get_async_session)
+    session: AsyncSession =Depends(get_async_session),
+    user: User = Depends(current_active_user)
 ):
     result = await session.execute(select(Post).order_by(Post.created_at.desc()))
     posts = [row[0] for row in result.all()]
+    result = await session.execute(select(User))
+    users = [row[0] for row in result.all()]
+    user_dict = {u.id:u.email for u in users}
     
     posts_data = []
     for post in posts:
         posts_data.append(
             {
                 "id": str(post.id),
+                "user_id": str(post.user.id),
                 "caption": post.caption,
                 "url":post.url,
                 "file_type": post.file_type,
                 "file_name": post.file_name,
-                "created_at":post.created_at.isoformat()
+                "created_at":post.created_at.isoformat(),
+                "is_owner": post.user_id == user.id,
+                "email":user_dict.get(post.user_id, "Unknown")
             }
         )
     
     return {"posts":posts_data}
 
 @app.delete("/posts/{post_id}")
-async def delete_post(post_id :str, session : AsyncSession = Depends(get_async_session)):
+async def delete_post(post_id :str, session : AsyncSession = Depends(get_async_session),
+                      user: User=Depends(current_active_user)):
     try:
         post_uuid = uuid.UUID(post_id)
         
@@ -95,6 +112,8 @@ async def delete_post(post_id :str, session : AsyncSession = Depends(get_async_s
         if not post:
             return HTTPException(status_code=404, detail="Post not Found")
         
+        if post.user_id != user.id:
+            raise HTTPException(status_code=403, detail="You don't have permission to delete this task")
         await session.delete(post)
         await session.commit()
